+++ docs/Changelog.md
@@ -27,6 +27,7 @@
   * When `"additionalProperties": T` type is `Map[String, T]`
 * [#3709](https://github.com/TouK/nussknacker/pull/3709) Fix BestEffortJsonSchemaEncoder - fix encoding only strings when `"type": String`
 * [#3707](https://github.com/TouK/nussknacker/pull/3707) Handling apikey in OpenAPI
+* [#3716](https://github.com/TouK/nussknacker/pull/3716) Fix: JsonSchema sinks allow to add additional fields also in `strict validation mode`, if schema permits them.
 
 1.6.1 (8 Nov 2022)
 ------------------------